extract code to utils
